+++ tests/test_http.py
@@ -606,17 +606,17 @@ class TestRange:
         assert rv.date is None
         assert rv.to_header() == '"Test"'
 
-        # weak information is dropped
+        # weak etag is discarded
         rv = IfRange.from_header('W/"Test"')
-        assert rv.etag == "Test"
+        assert rv.etag is None
         assert rv.date is None
-        assert rv.to_header() == '"Test"'
+        assert rv.to_header() == ""
 
         # broken etags are supported too
-        rv = IfRange.from_header("bullshit")
-        assert rv.etag == "bullshit"
+        rv = IfRange.from_header("unquoted")
+        assert rv.etag == "unquoted"
         assert rv.date is None
-        assert rv.to_header() == '"bullshit"'
+        assert rv.to_header() == '"unquoted"'
 
         rv = IfRange.from_header("Thu, 01 Jan 1970 00:00:00 GMT")
         assert rv.etag is None